vision_stock_analysis/requirements.txt
yfinance 
pandas 
plotly 
google-generativeai
kaleido
#pip install -U kaleido